requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.0
asgiref==3.7.2
asn1crypto==1.4.0
async-timeout==4.0.3
attrs==23.1.0
av==10.0.0
bcrypt==3.2.0
blinker==1.4
cached-property==1.5.2
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
colorama==0.4.4
coverage==7.2.7
croniter==1.4.1
cryptography==39.0.1
Cython==0.29.34
deprecation==2.1.0
distlib==0.3.4
distro==1.7.0
Django==4.2.3
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.100.0
fasteners==0.14.1
filelock==3.6.0
frozenlist==1.4.0
future==0.18.2
grpcio==1.58.0
grpcio-tools==1.58.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
idna==2.10
importlib-metadata==6.8.0
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
lagom==2.4.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
loguru==0.7.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mysql-connector-python==8.0.33
netifaces==0.11.0
numpy
oauthlib==3.2.0
olefile==0.46
packaging==23.1
paramiko==2.9.3
pexpect==4.8.0
pika==1.3.2
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.2.0
protobuf==3.20.3
psutil==5.9.0
psycopg==3.1.11
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.11.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-django==4.5.2
pytest-order==1.1.0
python-dateutil==2.8.1
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==3.0.0
python-multipart==0.0.6
pytz==2022.1
pyxdg==0.27
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
ssh-import-id==5.11
sshuttle==1.0.5
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.2
urllib3==1.26.5
uvicorn==0.23.0
wadllib==1.3.6
webencodings==0.5.1
xdg==5.0.0
yarl==1.9.2
zipp==1.0.0
email-validator==2.0.0.post2